🐛 fix: クラス再構成
